Move events from Admin to Library
