+++ src/libraries/kunena/src/Access/KunenaAccess.php
@@ -22,7 +22,7 @@ use Joomla\CMS\Language\Text;
 use Joomla\CMS\Plugin\PluginHelper;
 use Joomla\Database\Exception\ExecutionFailureException;
 use Joomla\Utilities\ArrayHelper;
-use Kunena\Forum\Administrator\Event\KunenaGetAccessControlEvent;
+use Kunena\Forum\Libraries\Event\KunenaGetAccessControlEvent;
 use Kunena\Forum\Libraries\Config\KunenaConfig;
 use Kunena\Forum\Libraries\Error\KunenaError;
 use Kunena\Forum\Libraries\Factory\KunenaFactory;
@@ -106,7 +106,7 @@ class KunenaAccess
     public function __construct()
     {
         KunenaProfiler::getInstance() ? KunenaProfiler::instance()->start('function ' . __CLASS__ . '::' . __FUNCTION__ . '()') : null;
-        PluginHelper::importPlugin('kunena','joomla');
+        PluginHelper::importPlugin('kunena', 'joomla');
 
         $accessControlEvent = new KunenaGetAccessControlEvent('onKunenaGetAccessControl');
         Factory::getApplication()->getDispatcher()->dispatch('onKunenaGetAccessControl', $accessControlEvent);
