+++ src/libraries/kunena/src/Event/KunenaGetAccessControlEvent.php
@@ -11,7 +11,7 @@
  * @link            https://www.kunena.org
  **/
 
-namespace Kunena\Forum\Administrator\Event;
+namespace Kunena\Forum\Libraries\Event;
 
 \defined('_JEXEC') or die;
 
@@ -20,15 +20,16 @@ use Joomla\CMS\Event\Result\ResultAware;
 use Joomla\CMS\Event\Result\ResultAwareInterface;
 use Joomla\CMS\Event\Result\ResultTypeObjectAware;
 use Kunena\Forum\Libraries\Integration\KunenaAvatar;
+use Kunena\Forum\Plugin\Kunena\Joomla\KunenaAccessJoomla;
 
 /**
- * Class for onKunenaGetAvatar event.
+ * Class for onKunenaGetAccessControl event.
  * Example:
- *  new KunenaGetAvatarEvent('onKunenaGetAvatar', []);
+ *  new KunenaGetAccessControlEvent('onKunenaGetAccessControl', []);
  *
  * @since  6.5.0
  */
-class KunenaGetAvatarEvent extends AbstractImmutableEvent implements ResultAwareInterface
+class KunenaGetAccessControlEvent extends AbstractImmutableEvent implements ResultAwareInterface
 {
     use ResultAware;
     use ResultTypeObjectAware;
@@ -42,10 +43,10 @@ class KunenaGetAvatarEvent extends AbstractImmutableEvent implements ResultAware
      * @since   6.5.0
      * @throws  \BadMethodCallException
      */
-    protected function onSetAvatar(KunenaAvatar $avatar): KunenaAvatar
+    /*protected function onSetAvatar(KunenaAvatar $avatar): KunenaAvatar
     {
         return $avatar;
-    }
+    }*/
 
     /**
      * Getter for the data argument.
@@ -53,12 +54,8 @@ class KunenaGetAvatarEvent extends AbstractImmutableEvent implements ResultAware
      * @return  KunenaAvatar|array
      * @since   6.5.0
      */
-    public function getAvatar(): KunenaAvatar|array
+    public function getAccessControl(): array
     {
-        if (isset($this->arguments['avatar']) && $this->arguments['avatar'] instanceof KunenaAvatar) {
-            return $this->arguments['avatar'];
-        }
-
         return $this->arguments['result'] ?? [];
     }
 
@@ -68,8 +65,8 @@ class KunenaGetAvatarEvent extends AbstractImmutableEvent implements ResultAware
      * @return  $this
      * @since   6.5.0
      */
-    public function setAvatar(KunenaAvatar $avatar)
+    /*public function setAvatar(KunenaAvatar $avatar)
     {
         return $this->setArgument('avatar', $avatar);
-    }
+    }*/
 }
