+++ src/libraries/kunena/src/Event/KunenaGetPrivateEvent.php
@@ -11,7 +11,7 @@
  * @link            https://www.kunena.org
  **/
 
-namespace Kunena\Forum\Administrator\Event;
+namespace Kunena\Forum\Libraries\Event;
 
 \defined('_JEXEC') or die;
 
@@ -19,16 +19,16 @@ use Joomla\CMS\Event\AbstractImmutableEvent;
 use Joomla\CMS\Event\Result\ResultAware;
 use Joomla\CMS\Event\Result\ResultAwareInterface;
 use Joomla\CMS\Event\Result\ResultTypeObjectAware;
-use Kunena\Forum\Libraries\Integration\KunenaProfile;
+use Kunena\Forum\Libraries\Integration\KunenaPrivate;
 
 /**
- * Class for onKunenaGetLogin event.
+ * Class for onKunenaGetPrivate event.
  * Example:
- *  new KunenaGetLoginEvent('onKunenaGetLogin', []);
+ *  new KunenaGetPrivateEvent('onKunenaGetPrivate', []);
  *
  * @since  6.5.0
  */
-class KunenaGetLoginEvent extends AbstractImmutableEvent implements ResultAwareInterface
+class KunenaGetPrivateEvent extends AbstractImmutableEvent implements ResultAwareInterface
 {
     use ResultAware;
     use ResultTypeObjectAware;
@@ -36,25 +36,29 @@ class KunenaGetLoginEvent extends AbstractImmutableEvent implements ResultAwareI
     /**
      * Setter for the object argument.
      *
-     * @param   KunenaProfile  $value  The value to set
+     * @param   KunenaPrivate  $value  The value to set
      *
-     * @return  KunenaProfile
+     * @return  KunenaPrivate
      * @since   6.5.0
      * @throws  \BadMethodCallException
      */
-    /*protected function onSetProfile(KunenaProfile $profile): KunenaProfile
+    protected function onSetPrivate(KunenaPrivate $private): KunenaPrivate
     {
-        return $profile;
-    }*/
+        return $private;
+    }
 
     /**
      * Getter for the data argument.
      *
-     * @return  array
+     * @return  KunenaPrivate|array
      * @since   6.5.0
      */
-    public function getLogin(): array
+    public function getPrivate(): KunenaPrivate|array
     {
+        if (isset($this->arguments['private']) && $this->arguments['private'] instanceof KunenaPrivate) {
+            return $this->arguments['private'];
+        }
+
         return $this->arguments['result'] ?? [];
     }
 
@@ -64,8 +68,8 @@ class KunenaGetLoginEvent extends AbstractImmutableEvent implements ResultAwareI
      * @return  $this
      * @since   6.5.0
      */
-    /*public function setProfile(KunenaProfile $profile)
+    public function setPrivate(KunenaPrivate $private)
     {
-        return $this->setArgument('profile', $profile);
-    }*/
+        return $this->setArgument('private', $private);
+    }
 }
