+++ src/plugins/plg_kunena_kunena/src/Extension/Kunena.php
@@ -20,8 +20,8 @@ use Joomla\CMS\Plugin\CMSPlugin;
 use Joomla\Database\DatabaseAwareInterface;
 use Joomla\Database\DatabaseAwareTrait;
 use Joomla\Event\SubscriberInterface;
-use Kunena\Forum\Administrator\Event\KunenaGetAvatarEvent;
-use Kunena\Forum\Administrator\Event\KunenaGetProfileEvent;
+use Kunena\Forum\Libraries\Event\KunenaGetAvatarEvent;
+use Kunena\Forum\Libraries\Event\KunenaGetProfileEvent;
 use Kunena\Forum\Plugin\Kunena\Kunena\Helper\KunenaAvatarKunena;
 use Kunena\Forum\Plugin\Kunena\Kunena\Helper\KunenaProfileKunena;
